requirements.txt
streamlit
opencv-python
google-generativeai
python-dotenv